Now can enter a transaction with decimal

+++ app/src/main/java/com/simoale/debitcredit/database/WalletDAO.java
@@ -22,5 +22,5 @@ public interface WalletDAO {
 
     @Transaction
     @Query("UPDATE wallet SET wallet_balance =  wallet_balance + :amount WHERE wallet_id = :walletId")
-    void updateBalance(Integer walletId, Integer amount);
+    void updateBalance(Integer walletId, Float amount);
 }
